Files used in September 2021 submission
These files were used to generate the data and figures in Sep'21 submission to HEALTH. Real data is ACTfast and Synthetic is either multivariate or 2-cluster gaussian. Perm in file name refers to the case when the number of features in two datasets is equal. Partial mapping is the ccase when there are some features in one dataset that have no match in another dataset. Onto case is when the datasets have different number of features but one of them is the subset of other. Sq Tran is the case when some of the features were square transformed in one dataset but the corresponding features in the other dataset are not.